Move token update and token refresh to repository.
