+++ pkg/service/api/token/token-service.go
@@ -2,19 +2,14 @@ package token
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"log/slog"
-	"slices"
 
 	"github.com/metal-stack/api/go/errorutil"
 	adminv2 "github.com/metal-stack/api/go/metalstack/admin/v2"
 	apiv2 "github.com/metal-stack/api/go/metalstack/api/v2"
 	"github.com/metal-stack/api/go/metalstack/api/v2/apiv2connect"
-	"github.com/metal-stack/api/go/permissions"
 	"github.com/metal-stack/metal-apiserver/pkg/repository/api"
 	"github.com/metal-stack/metal-apiserver/pkg/request"
-	"github.com/samber/lo"
 
 	"github.com/metal-stack/metal-apiserver/pkg/certs"
 	"github.com/metal-stack/metal-apiserver/pkg/repository"
@@ -95,74 +90,13 @@ func (t *tokenService) Update(ctx context.Context, req *apiv2.TokenServiceUpdate
 		return nil, errorutil.Unauthenticated("no token found in request")
 	}
 
-	// we first validate token permission elevation for the token used in the token update request,
-	// which might be an API token with restricted permissions
-
-	err := t.validateTokenRequest(ctx, token, req)
-	if err != nil {
-		return nil, errorutil.NewPermissionDenied(err)
-	}
-
-	// now, we validate if the user is still permitted to update the token
-	// doing this check is not strictly necessary because the resulting token would fail in the auther when being compared
-	// to the actual user permissions, but it's nicer for the user to already prevent token update immediately in this place
-
-	projectsAndTenants, err := t.projectsAndTenantsGetter(ctx, token.GetUser())
-	if err != nil {
-		return nil, errorutil.NewInternal(err)
-	}
-	fullUserToken := &apiv2.Token{
-		User:         token.User,
-		ProjectRoles: projectsAndTenants.ProjectRoles,
-		TenantRoles:  projectsAndTenants.TenantRoles,
-		AdminRole:    nil,
-		InfraRole:    token.InfraRole,
-		MachineRoles: token.MachineRoles,
-	}
-
-	if role, ok := t.hasAdminRole(projectsAndTenants); ok {
-		fullUserToken.AdminRole = role
-	}
-
-	err = t.validateTokenRequest(ctx, fullUserToken, req)
-	if err != nil {
-		return nil, errorutil.NewPermissionDenied(err)
-	}
-
-	// now follows the update
-
-	tokenToUpdate, err := t.tokens.Get(ctx, token.User, req.Uuid)
-	if err != nil {
-		return nil, err
-	}
-
-	if tokenToUpdate.TokenType != apiv2.TokenType_TOKEN_TYPE_API {
-		return nil, errorutil.FailedPrecondition("only updating API tokens is currently supported")
-	}
-
-	if req.Description != nil {
-		tokenToUpdate.Description = *req.Description
-	}
-	if req.AdminRole != nil {
-		if *req.AdminRole == apiv2.AdminRole_ADMIN_ROLE_UNSPECIFIED {
-			tokenToUpdate.AdminRole = nil
-		} else {
-			tokenToUpdate.AdminRole = req.AdminRole
-		}
-	}
-
-	tokenToUpdate.Permissions = req.Permissions
-	tokenToUpdate.ProjectRoles = req.ProjectRoles
-	tokenToUpdate.TenantRoles = req.TenantRoles
-	tokenToUpdate.MachineRoles = req.MachineRoles
-
-	err = t.tokens.Set(ctx, tokenToUpdate)
+	res, err := t.repo.Token(token.User).Update(ctx, req.Uuid, req)
 	if err != nil {
 		return nil, err
 	}
 
 	return &apiv2.TokenServiceUpdateResponse{
-		Token: tokenToUpdate,
+		Token: res.Token,
 	}, nil
 }
 
@@ -231,229 +165,5 @@ func (t *tokenService) Refresh(ctx context.Context, _ *apiv2.TokenServiceRefresh
 		return nil, errorutil.Unauthenticated("no token found in request")
 	}
 
-	oldtoken, err := t.tokens.Get(ctx, token.User, token.Uuid)
-	if err != nil {
-		return nil, err
-	}
-
-	// we first copy the token permission from the old token
-	createRequest := &apiv2.TokenServiceCreateRequest{
-		Permissions:  oldtoken.Permissions,
-		ProjectRoles: oldtoken.ProjectRoles,
-		TenantRoles:  oldtoken.TenantRoles,
-		AdminRole:    oldtoken.AdminRole,
-		InfraRole:    oldtoken.InfraRole,
-		MachineRoles: oldtoken.MachineRoles,
-	}
-
-	err = t.validateTokenRequest(ctx, token, createRequest)
-	if err != nil {
-		return nil, errorutil.NewPermissionDenied(err)
-	}
-
-	// now, we validate if the user is still permitted to refresh the token
-	// doing this check is not strictly necessary because the resulting token would fail in the auther when being compared
-	// to the actual user permissions, but it's nicer for the user to already prevent token update immediately in this place
-
-	projectsAndTenants, err := t.projectsAndTenantsGetter(ctx, token.GetUser())
-	if err != nil {
-		return nil, errorutil.NewInternal(err)
-	}
-	fullUserToken := &apiv2.Token{
-		User:         token.User,
-		ProjectRoles: projectsAndTenants.ProjectRoles,
-		TenantRoles:  projectsAndTenants.TenantRoles,
-		AdminRole:    nil,
-		InfraRole:    token.InfraRole,
-		MachineRoles: token.MachineRoles,
-	}
-	if role, ok := t.hasAdminRole(projectsAndTenants); ok {
-		fullUserToken.AdminRole = role
-	}
-	err = t.validateTokenRequest(ctx, fullUserToken, createRequest)
-	if err != nil {
-		return nil, errorutil.NewPermissionDenied(err)
-	}
-
-	// now follows the refresh, aka create a new token
-
-	privateKey, err := t.certs.LatestPrivate(ctx)
-	if err != nil {
-		return nil, errorutil.NewInternal(err)
-	}
-
-	// New duration is calculated from the old token
-	exp := oldtoken.Expires.AsTime().Sub(oldtoken.IssuedAt.AsTime())
-
-	secret, newToken, err := tokenutil.NewJWT(apiv2.TokenType_TOKEN_TYPE_API, token.GetUser(), t.issuer, exp, privateKey)
-	if err != nil {
-		return nil, errorutil.NewInternal(err)
-	}
-
-	newToken.Description = oldtoken.Description
-	newToken.Permissions = oldtoken.Permissions
-	newToken.ProjectRoles = oldtoken.ProjectRoles
-	newToken.TenantRoles = oldtoken.TenantRoles
-	newToken.AdminRole = oldtoken.AdminRole
-	newToken.InfraRole = oldtoken.InfraRole
-	newToken.MachineRoles = oldtoken.MachineRoles
-
-	err = t.tokens.Set(ctx, newToken)
-	if err != nil {
-		return nil, err
-	}
-
-	return &apiv2.TokenServiceRefreshResponse{
-		Token:  newToken,
-		Secret: secret,
-	}, nil
-}
-
-type tokenRequest interface {
-	GetPermissions() []*apiv2.MethodPermission
-	GetProjectRoles() map[string]apiv2.ProjectRole
-	GetTenantRoles() map[string]apiv2.TenantRole
-	GetMachineRoles() map[string]apiv2.MachineRole
-	GetAdminRole() apiv2.AdminRole
-	GetInfraRole() apiv2.InfraRole
-}
-
-func (t *tokenService) validateTokenRequest(ctx context.Context, currentToken *apiv2.Token, req tokenRequest) error {
-	// Calculate the permission from the token in the request
-	currentPermissions, err := t.authorizer.TokenPermissions(ctx, currentToken)
-	if err != nil {
-		return err
-	}
-
-	var (
-		adminRole *apiv2.AdminRole
-		infraRole *apiv2.InfraRole
-	)
-
-	// Ensure no unspecified roles are requested.
-	if req.GetAdminRole() != apiv2.AdminRole_ADMIN_ROLE_UNSPECIFIED {
-		adminRole = req.GetAdminRole().Enum()
-	}
-	if req.GetInfraRole() != apiv2.InfraRole_INFRA_ROLE_UNSPECIFIED {
-		infraRole = req.GetInfraRole().Enum()
-	}
-
-	var (
-		requestedTenants    = lo.Keys(req.GetTenantRoles())
-		allowedTenants      = lo.Keys(currentToken.TenantRoles)
-		forbiddenTenants, _ = lo.Difference(requestedTenants, allowedTenants)
-	)
-
-	if len(forbiddenTenants) > 0 && !slices.Contains(allowedTenants, "*") {
-		return fmt.Errorf("requested tenant roles are not allowed: %v", forbiddenTenants)
-	}
-
-	for _, tr := range req.GetTenantRoles() {
-		if tr == apiv2.TenantRole_TENANT_ROLE_UNSPECIFIED {
-			return fmt.Errorf("requested tenant role: %q is not allowed", tr)
-		}
-	}
-
-	var (
-		requestedProjects    = lo.Keys(req.GetProjectRoles())
-		allowedProjects      = lo.Keys(currentToken.ProjectRoles)
-		forbiddenProjects, _ = lo.Difference(requestedProjects, allowedProjects)
-	)
-
-	if len(forbiddenProjects) > 0 && !slices.Contains(allowedProjects, "*") {
-		return fmt.Errorf("requested project roles are not allowed: %v", forbiddenProjects)
-	}
-
-	for _, pr := range req.GetProjectRoles() {
-		if pr == apiv2.ProjectRole_PROJECT_ROLE_UNSPECIFIED {
-			return fmt.Errorf("requested project role: %q is not allowed", pr)
-		}
-	}
-
-	// Ensure no permissions pointing to unknown methods are requested
-	for _, permission := range req.GetPermissions() {
-		for _, method := range permission.Methods {
-			if _, found := permissions.GetServicePermissions().Methods[method]; !found {
-				return fmt.Errorf("unknown method %q", method)
-			}
-		}
-	}
-
-	var (
-		requestedMachines    = lo.Keys(req.GetMachineRoles())
-		allowedMachines      = lo.Keys(currentToken.MachineRoles)
-		forbiddenMachines, _ = lo.Difference(requestedMachines, allowedMachines)
-	)
-
-	if len(forbiddenMachines) > 0 && !slices.Contains(allowedMachines, "*") {
-		return fmt.Errorf("requested machine roles are not allowed: %v", forbiddenMachines)
-	}
-
-	// Calculate the permission from the token request (either create/update or refresh)
-	// and the methods which are coming from roles only.
-	requestedPermissions, err := t.authorizer.TokenPermissions(ctx, &apiv2.Token{
-		User:         currentToken.User,
-		Permissions:  req.GetPermissions(),
-		ProjectRoles: req.GetProjectRoles(),
-		TenantRoles:  req.GetTenantRoles(),
-		MachineRoles: req.GetMachineRoles(),
-		AdminRole:    adminRole,
-		InfraRole:    infraRole,
-	})
-	if err != nil {
-		return err
-	}
-
-	// sort requestedPermissions by method
-	var methods []string
-	for method := range requestedPermissions {
-		methods = append(methods, method)
-	}
-	slices.Sort(methods)
-
-	for _, method := range methods {
-		subjects, ok := requestedPermissions[method]
-		if !ok {
-			continue
-		}
-		currentSubjects, ok := currentPermissions[method]
-		if !ok {
-			errMsg := fmt.Sprintf("the following method %q is not allowed", method)
-			if len(subjects) > 0 {
-				errMsg += fmt.Sprintf(" on any of the requested subjects: %s", subjects)
-			}
-			return errors.New(errMsg)
-		}
-
-		if _, ok := currentSubjects[request.AnySubject]; ok {
-			continue
-		}
-		// It is possible to request any subjects to be able to have a token
-		// which is able to make calls to projects which will be created in the future.
-		// The actually possible subjects are calculated at request time.
-		if _, ok := subjects[request.AnySubject]; ok {
-			continue
-		}
-
-		for subject := range subjects {
-			if _, ok := currentSubjects[subject]; !ok {
-				return fmt.Errorf("method %q is not allowed on subject %q with your current user permissions", method, subject)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (t *tokenService) hasAdminRole(projectsAndTenants *api.ProjectsAndTenants) (*apiv2.AdminRole, bool) {
-
-	if role, ok := projectsAndTenants.TenantRoles[t.providerTenant]; ok {
-		switch role {
-		case apiv2.TenantRole_TENANT_ROLE_OWNER:
-			return apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(), true
-		case apiv2.TenantRole_TENANT_ROLE_EDITOR, apiv2.TenantRole_TENANT_ROLE_VIEWER:
-			return apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(), true
-		}
-	}
-	return nil, false
+	return t.repo.Token(token.User).AdditionalMethods().Refresh(ctx, token.Uuid)
 }
