+++ pkg/service/api/token/token-service_test.go
@@ -1,4 +1,4 @@
-package token
+package token_test
 
 import (
 	"context"
@@ -7,16 +7,17 @@ import (
 	"time"
 
 	"buf.build/go/protovalidate"
-	"github.com/alicebob/miniredis/v2"
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+	"github.com/metal-stack/api/go/errorutil"
 	apiv2 "github.com/metal-stack/api/go/metalstack/api/v2"
-	"github.com/metal-stack/metal-apiserver/pkg/certs"
 	"github.com/metal-stack/metal-apiserver/pkg/repository/api"
-	"github.com/metal-stack/metal-apiserver/pkg/request"
+	tokenservice "github.com/metal-stack/metal-apiserver/pkg/service/api/token"
+	"github.com/metal-stack/metal-apiserver/pkg/test"
 	"github.com/metal-stack/metal-apiserver/pkg/token"
-	"github.com/redis/go-redis/v9"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/testing/protocmp"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
@@ -25,8 +26,1464 @@ var (
 	token1 = "00000000-0000-0000-0000-000000000000"
 )
 
+func Test_Create(t *testing.T) {
+	t.Parallel()
+
+	log := slog.Default()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log, test.WithValkey(true), test.WithPostgres(true))
+	defer closer()
+
+	type state struct {
+		providerTenant string
+		projectRoles   map[string]apiv2.ProjectRole
+		tenantRoles    map[string]apiv2.TenantRole
+	}
+	tests := []struct {
+		name           string
+		sessionToken   *apiv2.Token
+		req            *apiv2.TokenServiceCreateRequest
+		state          state
+		wantErr        bool
+		wantErrMessage string
+		wantToken      *apiv2.Token
+	}{
+		{
+			name: "can create bare token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "empty token",
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles:    map[string]apiv2.TenantRole{},
+			},
+			wantToken: &apiv2.Token{
+				User:        "phippy",
+				Description: "empty token",
+				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+			},
+		},
+		{
+			name: "user and token without project access cannot create project token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested project roles are not allowed: [00000000-0000-0000-0000-000000000000]`,
+		},
+		{
+			name: "user and token with project access can create project token",
+			sessionToken: &apiv2.Token{
+				User:        "phippy",
+				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions: []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				projectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+			},
+			wantToken: &apiv2.Token{
+				User:        "phippy",
+				Description: "project token",
+				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+		},
+		{
+			name: "user without but token with project access cannot create project token",
+			sessionToken: &apiv2.Token{
+				User:        "phippy",
+				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions: []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				projectRoles:   map[string]apiv2.ProjectRole{},
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested project roles are not allowed: [00000000-0000-0000-0000-000000000000]`,
+		},
+		{
+			name: "project without but user with project access cannot create project token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				projectRoles: map[string]apiv2.ProjectRole{
+					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
+				},
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested project roles are not allowed: [00000000-0000-0000-0000-000000000000]`,
+		},
+		{
+			name: "normal user with provider-tenant owner membership can create new admin editor token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_USER,
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles: map[string]apiv2.TenantRole{
+					test.DefaultProviderTenant: apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			},
+			wantToken: &apiv2.Token{
+				User:         "phippy",
+				Description:  "admin token",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+		},
+		{
+			name: "normal user with provider-tenant viewer membership can create new admin viewer token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_USER,
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles: map[string]apiv2.TenantRole{
+					test.DefaultProviderTenant: apiv2.TenantRole_TENANT_ROLE_VIEWER,
+				},
+			},
+			wantToken: &apiv2.Token{
+				User:         "phippy",
+				Description:  "admin token",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(),
+			},
+		},
+		{
+			name: "provider tenant user owner can create new admin editor token",
+			sessionToken: &apiv2.Token{
+				User:         test.DefaultProviderTenant,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_USER,
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantToken: &apiv2.Token{
+				User:         test.DefaultProviderTenant,
+				Description:  "admin token",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+		},
+		{
+			name: "provider tenant user viewer can create new admin viewer token",
+			sessionToken: &apiv2.Token{
+				User:         test.DefaultProviderTenant,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_USER,
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles: map[string]apiv2.TenantRole{
+					test.DefaultProviderTenant: apiv2.TenantRole_TENANT_ROLE_VIEWER,
+				},
+			},
+			wantToken: &apiv2.Token{
+				User:         test.DefaultProviderTenant,
+				Description:  "admin token",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(),
+			},
+		},
+		{
+			name: "admin viewer cannot create admin editor token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_USER,
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles: map[string]apiv2.TenantRole{
+					test.DefaultProviderTenant: apiv2.TenantRole_TENANT_ROLE_VIEWER,
+				},
+			},
+			wantToken:      nil,
+			wantErr:        true,
+			wantErrMessage: `permission_denied: your provider tenant membership only allows "ADMIN_ROLE_VIEWER", but you requested "ADMIN_ROLE_EDITOR"`,
+		},
+		{
+			name: "normal user which is not member in the provider tenant can not create new admin viewer token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_USER,
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_VIEWER.Enum(),
+			},
+			state: state{
+				providerTenant: "blippy",
+			},
+			wantToken:      nil,
+			wantErr:        true,
+			wantErrMessage: `permission_denied: the following method "/metalstack.admin.v2.AuditService/Get" is not allowed on any of the requested subjects: [*]`,
+		},
+		{
+			name: "admin user and token can create new admin token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles: map[string]apiv2.TenantRole{
+					test.DefaultProviderTenant: apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			},
+			wantToken: &apiv2.Token{
+				User:         "phippy",
+				Description:  "admin token",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+		},
+		{
+			name: "admin token but not user cannot create new admin token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "admin token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: the following method "/grpc.reflection.v1.ServerReflection/ServerReflectionInfo" is not allowed on any of the requested subjects: [*]`,
+		},
+		{
+			name: "user and token without tenant access cannot create tenant token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested tenant roles are not allowed: [mascots]`,
+		},
+		{
+			name: "user and token with tenant access can create tenant token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				tenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			wantToken: &apiv2.Token{
+				User:         "phippy",
+				Description:  "project token",
+				TokenType:    *apiv2.TokenType_TOKEN_TYPE_API.Enum(),
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+		},
+		{
+			name: "user requests token for mascots but in the database does not have required tenant membership for mascots",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				projectRoles:   map[string]apiv2.ProjectRole{},
+				tenantRoles: map[string]apiv2.TenantRole{
+					"phippy": apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested tenant roles are not allowed: [mascots]`,
+		},
+		{
+			name: "user requests token for mascots but neither has mascots in his token permissions nor in the database",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"phippy": apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				projectRoles:   map[string]apiv2.ProjectRole{},
+				tenantRoles: map[string]apiv2.TenantRole{
+					"phippy": apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested tenant roles are not allowed: [mascots]`,
+		},
+		{
+			name: "token without but user with tenant access cannot create tenant token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description:  "project token",
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+				projectRoles:   map[string]apiv2.ProjectRole{},
+				tenantRoles: map[string]apiv2.TenantRole{
+					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
+				},
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested tenant roles are not allowed: [mascots]`,
+		},
+		{
+			name: "user and token without machine access cannot create machine token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "machine token",
+				MachineRoles: map[string]apiv2.MachineRole{
+					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantErr:        true,
+			wantErrMessage: `permission_denied: requested machine roles are not allowed: [de240964-ff9f-4e3d-95b2-8a96e43788f1]`,
+		},
+		{
+			name: "user and token with machine access can create machine token",
+			sessionToken: &apiv2.Token{
+				User:        "pixie-core",
+				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions: []*apiv2.MethodPermission{},
+				MachineRoles: map[string]apiv2.MachineRole{
+					"*": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceCreateRequest{
+				Description: "machine token",
+				MachineRoles: map[string]apiv2.MachineRole{
+					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantToken: &apiv2.Token{
+				User:        "pixie-core",
+				Description: "machine token",
+				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+				MachineRoles: map[string]apiv2.MachineRole{
+					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+				TenantRoles: map[string]apiv2.TenantRole{},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(innerT *testing.T) {
+			defer testStore.Cleanup(t)
+
+			ctx, cancel := context.WithCancel(token.ContextWithToken(innerT.Context(), tt.sessionToken))
+			defer cancel()
+
+			test.CreateTenants(innerT, testStore, []*apiv2.TenantServiceCreateRequest{
+				{
+					Name: tt.sessionToken.User,
+				},
+			})
+
+			// every logged in users comes with default tenant owner membership
+			test.CreateTenantMemberships(innerT, testStore, tt.sessionToken.User, []*api.TenantMemberCreateRequest{
+				{
+					MemberID: tt.sessionToken.User,
+					Role:     apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			})
+
+			for id, perm := range tt.state.tenantRoles {
+				if id != tt.sessionToken.User {
+					test.CreateTenants(innerT, testStore, []*apiv2.TenantServiceCreateRequest{
+						{
+							Name: id,
+						},
+					})
+				}
+				test.CreateTenantMemberships(innerT, testStore, id, []*api.TenantMemberCreateRequest{
+					{
+						MemberID: tt.sessionToken.User,
+						Role:     perm,
+					},
+				})
+			}
+
+			for id, perm := range tt.state.projectRoles {
+				test.CreateProjects(innerT, testStore, []*apiv2.ProjectServiceCreateRequest{
+					{
+						Login: tt.sessionToken.User,
+						Name:  id,
+					},
+				})
+				test.CreateProjectMemberships(innerT, testStore, id, []*api.ProjectMemberCreateRequest{
+					{
+						TenantId: tt.sessionToken.User,
+						Role:     perm,
+					},
+				})
+			}
+
+			service := tokenservice.New(tokenservice.Config{
+				Log:  log,
+				Repo: testStore.Store,
+			})
+
+			if tt.wantErr == false {
+				// Execute proto based validation
+				err := protovalidate.Validate(tt.req)
+				require.NoError(t, err)
+			}
+
+			response, err := service.Create(ctx, tt.req)
+			switch {
+			case tt.wantErr && err != nil:
+				if diff := cmp.Diff(tt.wantErrMessage, err.Error()); diff != "" {
+					t.Errorf("diff = %s", diff)
+				}
+			case tt.wantErr && err == nil:
+				t.Fatalf("want error %q, got response %q", tt.wantErrMessage, response)
+			case err != nil:
+				t.Fatalf("want response, got error %q", err)
+
+			default:
+				if response.Secret == "" {
+					t.Error("response secret for token may not be empty")
+				}
+				require.NotNil(t, tt.wantToken, "token returned, nil expected")
+
+				got := response.Token
+				assert.Equal(t, tt.wantToken.Description, got.Description, "description")
+				assert.Equal(t, tt.wantToken.User, got.User, "user id")
+				assert.Equal(t, tt.wantToken.TokenType, got.TokenType, "token type")
+				assert.Equal(t, tt.wantToken.AdminRole, got.AdminRole, "admin role")
+				assert.Equal(t, tt.wantToken.Permissions, got.Permissions, "permissions")
+				assert.Equal(t, tt.wantToken.ProjectRoles, got.ProjectRoles, "project roles")
+				assert.Equal(t, tt.wantToken.TenantRoles, got.TenantRoles, "tenant roles")
+				assert.Equal(t, tt.wantToken.MachineRoles, got.MachineRoles, "machine roles")
+			}
+		})
+	}
+}
+
+func Test_Create_NoToken(t *testing.T) {
+	t.Parallel()
+
+	log := slog.Default()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log, test.WithValkey(true), test.WithPostgres(true))
+	defer closer()
+
+	service := tokenservice.New(tokenservice.Config{
+		Log:            log,
+		Repo:           testStore.Store,
+		ProviderTenant: test.DefaultProviderTenant,
+		Issuer:         "http://test",
+	})
+
+	_, err := service.Create(t.Context(), &apiv2.TokenServiceCreateRequest{})
+	require.Error(t, err)
+	require.Equal(t, "unauthenticated: no token found in request", err.Error())
+}
+
+func Test_List(t *testing.T) {
+	t.Parallel()
+
+	log := slog.Default()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log)
+	defer closer()
+
+	type state struct {
+		existingTokens []*apiv2.Token
+	}
+	tests := []struct {
+		name           string
+		sessionToken   *apiv2.Token
+		req            *apiv2.TokenServiceListRequest
+		state          state
+		wantErr        bool
+		wantErrMessage string
+		want           []*apiv2.Token
+	}{
+		{
+			name: "no tokens",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req:   &apiv2.TokenServiceListRequest{},
+			state: state{},
+			want:  nil,
+		},
+		{
+			name: "list tokens",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceListRequest{},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+		},
+		{
+			name: "query uuid",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					Uuid: new("c223af4d-b3f5-4df6-8815-52b80323930d"),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+		},
+		{
+			name: "query description and labels",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					Labels: &apiv2.Labels{
+						Labels: map[string]string{
+							"a": "b",
+							"c": "d",
+						},
+					},
+					Description: new("test"),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid:        "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User:        "phippy",
+						Description: "test",
+						Meta: &apiv2.Meta{
+							Labels: &apiv2.Labels{
+								Labels: map[string]string{
+									"c": "d",
+									"a": "b",
+								},
+							},
+						},
+					},
+					{
+						Uuid:        "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User:        "phippy",
+						Description: "nope",
+						Meta: &apiv2.Meta{
+							Labels: &apiv2.Labels{
+								Labels: map[string]string{
+									"a": "b",
+									"c": "d",
+								},
+							},
+						},
+					},
+					{
+						Uuid:        "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User:        "phippy",
+						Description: "test",
+						Meta: &apiv2.Meta{
+							Labels: &apiv2.Labels{
+								Labels: map[string]string{
+									"a": "b",
+									"c": "nope",
+								},
+							},
+						},
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid:        "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User:        "phippy",
+					Description: "test",
+					Meta: &apiv2.Meta{
+						Labels: &apiv2.Labels{
+							Labels: map[string]string{
+								"a": "b",
+								"c": "d",
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "query user (does not see other users)",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					User: new("not phippy"),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: nil,
+		},
+		{
+			name: "query token type",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceListRequest{
+				Query: &apiv2.TokenQuery{
+					TokenType: new(apiv2.TokenType_TOKEN_TYPE_API),
+				},
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid:      "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User:      "phippy",
+						TokenType: apiv2.TokenType_TOKEN_TYPE_API,
+					},
+					{
+						Uuid:      "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User:      "phippy",
+						TokenType: apiv2.TokenType_TOKEN_TYPE_USER,
+					},
+				},
+			},
+			want: []*apiv2.Token{
+				{
+					Uuid:      "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User:      "phippy",
+					TokenType: apiv2.TokenType_TOKEN_TYPE_API,
+					Meta:      &apiv2.Meta{},
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(innerT *testing.T) {
+			defer testStore.Cleanup(t)
+
+			ctx, cancel := context.WithCancel(token.ContextWithToken(innerT.Context(), tt.sessionToken))
+			defer cancel()
+
+			for _, tok := range tt.state.existingTokens {
+				err := testStore.GetTokenStore().Set(ctx, tok)
+				require.NoError(innerT, err)
+			}
+
+			service := tokenservice.New(tokenservice.Config{
+				Log:  log,
+				Repo: testStore.Store,
+			})
+
+			if tt.wantErr == false {
+				// Execute proto based validation
+				err := protovalidate.Validate(tt.req)
+				require.NoError(innerT, err)
+			}
+
+			response, err := service.List(ctx, tt.req)
+
+			switch {
+			case tt.wantErr && err != nil:
+				if diff := cmp.Diff(tt.wantErrMessage, err.Error()); diff != "" {
+					innerT.Errorf("diff = %s", diff)
+				}
+
+			case tt.wantErr && err == nil:
+				innerT.Fatalf("want error %q, got response %q", tt.wantErrMessage, response)
+			case err != nil:
+				innerT.Fatalf("want response, got error %q", err)
+
+			default:
+				if diff := cmp.Diff(
+					tt.want, response.Tokens,
+					cmpopts.SortSlices(func(a, b *apiv2.Token) bool {
+						return a.Uuid < b.Uuid
+					}),
+					protocmp.Transform(),
+					protocmp.IgnoreFields(
+						&apiv2.Token{}, "issued_at", "expires",
+					),
+					protocmp.IgnoreFields(
+						&apiv2.Meta{}, "created_at", "updated_at",
+					),
+				); diff != "" {
+					innerT.Errorf("diff: %s", diff)
+				}
+			}
+		})
+	}
+}
+
+func Test_Get(t *testing.T) {
+	t.Parallel()
+
+	log := slog.Default()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log)
+	defer closer()
+
+	type state struct {
+		existingTokens []*apiv2.Token
+	}
+	tests := []struct {
+		name              string
+		sessionToken      *apiv2.Token
+		req               *apiv2.TokenServiceGetRequest
+		state             state
+		wantErr           error
+		wantValidationErr string
+		want              *apiv2.Token
+	}{
+		{
+			name: "missing id in request",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceGetRequest{},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want:              nil,
+			wantValidationErr: "validation error: uuid: value is empty, which is not a valid UUID",
+		},
+		{
+			name: "get non-existing",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceGetRequest{
+				Uuid: "57460ff2-30e9-45e5-93c8-7f9ca85a92af",
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			wantErr: errorutil.NotFound("token not found"),
+		},
+		{
+			name: "cannot get another user's token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceGetRequest{
+				Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			wantErr: errorutil.NotFound("token not found"),
+		},
+		{
+			name: "get existing",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceGetRequest{
+				Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			want: &apiv2.Token{
+				Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+				User: "phippy",
+				Meta: &apiv2.Meta{},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(innerT *testing.T) {
+			defer testStore.Cleanup(t)
+
+			ctx, cancel := context.WithCancel(token.ContextWithToken(innerT.Context(), tt.sessionToken))
+			defer cancel()
+
+			for _, tok := range tt.state.existingTokens {
+				err := testStore.GetTokenStore().Set(ctx, tok)
+				require.NoError(innerT, err)
+			}
+
+			service := tokenservice.New(tokenservice.Config{
+				Log:  log,
+				Repo: testStore.Store,
+			})
+
+			// Execute proto based validation
+			err := protovalidate.Validate(tt.req)
+			if tt.wantValidationErr != "" {
+				if diff := cmp.Diff(tt.wantValidationErr, err.Error(), errorutil.ErrorStringComparer()); diff != "" {
+					innerT.Errorf("diff = %s", diff)
+				}
+
+				return
+			}
+
+			require.NoError(innerT, err)
+
+			response, err := service.Get(ctx, tt.req)
+
+			switch {
+			case tt.wantErr != nil && err != nil:
+				if diff := cmp.Diff(tt.wantErr, err, errorutil.ErrorStringComparer()); diff != "" {
+					innerT.Errorf("diff = %s", diff)
+				}
+
+			case tt.wantErr != nil && err == nil:
+				innerT.Fatalf("want error %q, got response %q", tt.wantErr, response)
+			case err != nil:
+				innerT.Fatalf("want response, got error %q", err)
+
+			default:
+				if diff := cmp.Diff(
+					tt.want, response.Token,
+					protocmp.Transform(),
+					protocmp.IgnoreFields(
+						&apiv2.Token{}, "issued_at", "expires",
+					),
+					protocmp.IgnoreFields(
+						&apiv2.Meta{}, "created_at", "updated_at",
+					),
+				); diff != "" {
+					innerT.Errorf("diff: %s", diff)
+				}
+			}
+		})
+	}
+}
+
+func Test_Revoke(t *testing.T) {
+	t.Parallel()
+
+	log := slog.Default()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log)
+	defer closer()
+
+	type state struct {
+		existingTokens []*apiv2.Token
+	}
+	tests := []struct {
+		name              string
+		sessionToken      *apiv2.Token
+		req               *apiv2.TokenServiceRevokeRequest
+		state             state
+		wantErr           error
+		wantValidationErr string
+		wantRemaining     []*apiv2.Token
+	}{
+		{
+			name: "missing id in request",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceRevokeRequest{},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			wantRemaining: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+					User: "not phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+			wantValidationErr: "validation error: uuid: value is empty, which is not a valid UUID",
+		},
+		{
+			name: "revoke non-existing",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceRevokeRequest{
+				Uuid: "57460ff2-30e9-45e5-93c8-7f9ca85a92af",
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			wantRemaining: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+					User: "not phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+			wantErr: errorutil.NotFound("token not found"),
+		},
+		{
+			name: "cannot revoke another user's token",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceRevokeRequest{
+				Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			wantRemaining: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+					User: "not phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+			wantErr: errorutil.NotFound("token not found"),
+		},
+		{
+			name: "revoke existing",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+			},
+			req: &apiv2.TokenServiceRevokeRequest{
+				Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+			},
+			state: state{
+				existingTokens: []*apiv2.Token{
+					{
+						Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+						User: "phippy",
+					},
+					{
+						Uuid: "8ff27ee2-209f-43e2-a15d-50143fb03229",
+						User: "phippy",
+					},
+					{
+						Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+						User: "not phippy",
+					},
+				},
+			},
+			wantRemaining: []*apiv2.Token{
+				{
+					Uuid: "c223af4d-b3f5-4df6-8815-52b80323930d",
+					User: "phippy",
+					Meta: &apiv2.Meta{},
+				},
+				{
+					Uuid: "9baa8668-2212-4fa5-a2e4-167084d0552d",
+					User: "not phippy",
+					Meta: &apiv2.Meta{},
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(innerT *testing.T) {
+			defer testStore.Cleanup(t)
+
+			ctx, cancel := context.WithCancel(token.ContextWithToken(innerT.Context(), tt.sessionToken))
+			defer cancel()
+
+			for _, tok := range tt.state.existingTokens {
+				err := testStore.GetTokenStore().Set(ctx, tok)
+				require.NoError(innerT, err)
+			}
+
+			service := tokenservice.New(tokenservice.Config{
+				Log:  log,
+				Repo: testStore.Store,
+			})
+
+			// Execute proto based validation
+			err := protovalidate.Validate(tt.req)
+			if tt.wantValidationErr != "" {
+				if diff := cmp.Diff(tt.wantValidationErr, err.Error(), errorutil.ErrorStringComparer()); diff != "" {
+					innerT.Errorf("diff = %s", diff)
+				}
+
+				return
+			}
+
+			require.NoError(innerT, err)
+
+			response, err := service.Revoke(ctx, tt.req)
+
+			switch {
+			case tt.wantErr != nil && err != nil:
+				if diff := cmp.Diff(tt.wantErr, err, errorutil.ErrorStringComparer()); diff != "" {
+					innerT.Errorf("diff = %s", diff)
+				}
+
+			case tt.wantErr != nil && err == nil:
+				innerT.Fatalf("want error %q, got response %q", tt.wantErr, response)
+			case err != nil:
+				innerT.Fatalf("want response, got error %q", err)
+
+			default:
+				remaining, err := testStore.GetTokenStore().AdminList(ctx)
+				require.NoError(innerT, err)
+
+				if diff := cmp.Diff(
+					tt.wantRemaining, remaining,
+					protocmp.Transform(),
+					protocmp.IgnoreFields(
+						&apiv2.Token{}, "issued_at", "expires",
+					),
+					protocmp.IgnoreFields(
+						&apiv2.Meta{}, "created_at", "updated_at",
+					),
+					cmpopts.SortSlices(func(a, b *apiv2.Token) bool {
+						return a.Uuid < b.Uuid
+					}),
+				); diff != "" {
+					innerT.Errorf("diff: %s", diff)
+				}
+			}
+		})
+	}
+}
+
 func Test_Update(t *testing.T) {
 	t.Parallel()
+
+	log := slog.Default()
+	now := time.Now()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log, test.WithValkey(true), test.WithPostgres(true))
+	defer closer()
+
 	type state struct {
 		providerTenant string
 		projectRoles   map[string]apiv2.ProjectRole
@@ -51,6 +1508,9 @@ func Test_Update(t *testing.T) {
 				TenantRoles:  map[string]apiv2.TenantRole{},
 			},
 			tokenToUpdate: &apiv2.Token{
+				Meta: &apiv2.Meta{
+					CreatedAt: timestamppb.New(now),
+				},
 				Uuid:         token1,
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -61,15 +1521,36 @@ func Test_Update(t *testing.T) {
 			req: &apiv2.TokenServiceUpdateRequest{
 				Uuid:        token1,
 				Description: new("update!"),
+				Labels: &apiv2.UpdateLabels{
+					Strategy: &apiv2.UpdateLabels_Replace{
+						Replace: &apiv2.Labels{
+							Labels: map[string]string{
+								"a": "b",
+							},
+						},
+					},
+				},
+				UpdateMeta: &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantToken: &apiv2.Token{
-				Uuid:        token1,
-				User:        "phippy",
-				Description: "update!",
-				TokenType:   apiv2.TokenType_TOKEN_TYPE_API,
+				Uuid:         token1,
+				User:         "phippy",
+				Description:  "update!",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				MachineRoles: map[string]apiv2.MachineRole{},
+				Meta: &apiv2.Meta{
+					Generation: 1,
+					Labels: &apiv2.Labels{
+						Labels: map[string]string{
+							"a": "b",
+						},
+					},
+				},
 			},
 		},
 		{
@@ -93,9 +1574,10 @@ func Test_Update(t *testing.T) {
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
 				TenantRoles: map[string]apiv2.TenantRole{},
+				UpdateMeta:  &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantErr:        true,
 			wantErrMessage: `permission_denied: requested project roles are not allowed: [00000000-0000-0000-0000-000000000000]`,
@@ -124,9 +1606,10 @@ func Test_Update(t *testing.T) {
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
 				TenantRoles: map[string]apiv2.TenantRole{},
+				UpdateMeta:  &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles: map[string]apiv2.ProjectRole{
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
@@ -138,11 +1621,15 @@ func Test_Update(t *testing.T) {
 				ProjectRoles: map[string]apiv2.ProjectRole{
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
-				TenantRoles: map[string]apiv2.TenantRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				MachineRoles: map[string]apiv2.MachineRole{},
+				Meta: &apiv2.Meta{
+					Generation: 1,
+				},
 			},
 		},
 		{
-			name: "user without but token with project access cannot update project token",
+			name: "user with missing project membership in the database cannot update project token",
 			sessionToken: &apiv2.Token{
 				User:        "phippy",
 				Permissions: []*apiv2.MethodPermission{},
@@ -164,16 +1651,17 @@ func Test_Update(t *testing.T) {
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
 				TenantRoles: map[string]apiv2.TenantRole{},
+				UpdateMeta:  &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles:   map[string]apiv2.ProjectRole{},
 			},
 			wantErr:        true,
 			wantErrMessage: `permission_denied: requested project roles are not allowed: [00000000-0000-0000-0000-000000000000]`,
 		},
 		{
-			name: "project without but user with project access cannot create project token",
+			name: "token with missing project roles cannot create project token despite being member in the database",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -193,9 +1681,10 @@ func Test_Update(t *testing.T) {
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
 				TenantRoles: map[string]apiv2.TenantRole{},
+				UpdateMeta:  &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles: map[string]apiv2.ProjectRole{
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
@@ -204,7 +1693,7 @@ func Test_Update(t *testing.T) {
 			wantErrMessage: `permission_denied: requested project roles are not allowed: [00000000-0000-0000-0000-000000000000]`,
 		},
 		{
-			name: "admin user and token can update admin token",
+			name: "admin with provider tenant membership and admin token can update token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -226,11 +1715,12 @@ func Test_Update(t *testing.T) {
 				ProjectRoles: map[string]apiv2.ProjectRole{},
 				TenantRoles:  map[string]apiv2.TenantRole{},
 				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+				UpdateMeta:   &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "phippy",
+				providerTenant: test.DefaultProviderTenant,
 				tenantRoles: map[string]apiv2.TenantRole{
-					"phippy": apiv2.TenantRole_TENANT_ROLE_OWNER,
+					test.DefaultProviderTenant: apiv2.TenantRole_TENANT_ROLE_OWNER,
 				},
 			},
 			wantToken: &apiv2.Token{
@@ -240,10 +1730,14 @@ func Test_Update(t *testing.T) {
 				ProjectRoles: map[string]apiv2.ProjectRole{},
 				TenantRoles:  map[string]apiv2.TenantRole{},
 				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+				MachineRoles: map[string]apiv2.MachineRole{},
+				Meta: &apiv2.Meta{
+					Generation: 1,
+				},
 			},
 		},
 		{
-			name: "admin token but user cannot update admin token",
+			name: "admin without provider tenant membership cannot update admin token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -257,12 +1751,14 @@ func Test_Update(t *testing.T) {
 				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
 			},
 			req: &apiv2.TokenServiceUpdateRequest{
+				Uuid:         token1,
 				ProjectRoles: map[string]apiv2.ProjectRole{},
 				TenantRoles:  map[string]apiv2.TenantRole{},
 				AdminRole:    apiv2.AdminRole_ADMIN_ROLE_EDITOR.Enum(),
+				UpdateMeta:   &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantErr:        true,
 			wantErrMessage: `permission_denied: the following method "/grpc.reflection.v1.ServerReflection/ServerReflectionInfo" is not allowed on any of the requested subjects: [*]`,
@@ -286,9 +1782,10 @@ func Test_Update(t *testing.T) {
 				TenantRoles: map[string]apiv2.TenantRole{
 					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
 				},
+				UpdateMeta: &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantErr:        true,
 			wantErrMessage: `permission_denied: requested tenant roles are not allowed: [mascots]`,
@@ -317,9 +1814,10 @@ func Test_Update(t *testing.T) {
 				TenantRoles: map[string]apiv2.TenantRole{
 					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
 				},
+				UpdateMeta: &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				tenantRoles: map[string]apiv2.TenantRole{
 					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
 				},
@@ -332,10 +1830,14 @@ func Test_Update(t *testing.T) {
 				TenantRoles: map[string]apiv2.TenantRole{
 					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
 				},
+				MachineRoles: map[string]apiv2.MachineRole{},
+				Meta: &apiv2.Meta{
+					Generation: 1,
+				},
 			},
 		},
 		{
-			name: "user without but token with tenant access cannot update tenant token",
+			name: "user without tenant membership in the database cannot update tenant token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -357,14 +1859,14 @@ func Test_Update(t *testing.T) {
 				},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles:   map[string]apiv2.ProjectRole{},
 			},
 			wantErr:        true,
 			wantErrMessage: `permission_denied: requested tenant roles are not allowed: [mascots]`,
 		},
 		{
-			name: "user and token with machine access can update machine token",
+			name: "token with machine access can update machine token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -388,21 +1890,27 @@ func Test_Update(t *testing.T) {
 				MachineRoles: map[string]apiv2.MachineRole{
 					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
 				},
+				UpdateMeta: &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantToken: &apiv2.Token{
-				Uuid:      token1,
-				User:      "phippy",
-				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
+				Uuid:         token1,
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
 				MachineRoles: map[string]apiv2.MachineRole{
 					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+				Meta: &apiv2.Meta{
+					Generation: 1,
 				},
 			},
 		},
 		{
-			name: "user and token without machine access cannot update machine token",
+			name: "token without machine access cannot update machine token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -419,15 +1927,16 @@ func Test_Update(t *testing.T) {
 				MachineRoles: map[string]apiv2.MachineRole{
 					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
 				},
+				UpdateMeta: &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantErr:        true,
 			wantErrMessage: `permission_denied: requested machine roles are not allowed: [de240964-ff9f-4e3d-95b2-8a96e43788f1]`,
 		},
 		{
-			name: "token without but user with tenant access cannot update tenant token",
+			name: "token missing tenant role despite having tenant membership in the database cannot update tenant token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Permissions:  []*apiv2.MethodPermission{},
@@ -444,9 +1953,10 @@ func Test_Update(t *testing.T) {
 				TenantRoles: map[string]apiv2.TenantRole{
 					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
 				},
+				UpdateMeta: &apiv2.UpdateMeta{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles:   map[string]apiv2.ProjectRole{},
 				tenantRoles: map[string]apiv2.TenantRole{
 					"mascots": apiv2.TenantRole_TENANT_ROLE_EDITOR,
@@ -477,7 +1987,7 @@ func Test_Update(t *testing.T) {
 				TenantRoles: map[string]apiv2.TenantRole{},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles: map[string]apiv2.ProjectRole{
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
@@ -486,42 +1996,105 @@ func Test_Update(t *testing.T) {
 			wantErr:        true,
 			wantErrMessage: `not_found: token not found`,
 		},
+		{
+			name: "cannot demand optimistic locking for now",
+			sessionToken: &apiv2.Token{
+				User:         "phippy",
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				MachineRoles: map[string]apiv2.MachineRole{
+					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+			},
+			tokenToUpdate: &apiv2.Token{
+				Uuid:         token1,
+				User:         "phippy",
+				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
+				Permissions:  []*apiv2.MethodPermission{},
+				ProjectRoles: map[string]apiv2.ProjectRole{},
+				TenantRoles:  map[string]apiv2.TenantRole{},
+				MachineRoles: map[string]apiv2.MachineRole{},
+			},
+			req: &apiv2.TokenServiceUpdateRequest{
+				Uuid: token1,
+				MachineRoles: map[string]apiv2.MachineRole{
+					"de240964-ff9f-4e3d-95b2-8a96e43788f1": apiv2.MachineRole_MACHINE_ROLE_EDITOR,
+				},
+				UpdateMeta: &apiv2.UpdateMeta{
+					UpdatedAt:       timestamppb.Now(),
+					LockingStrategy: apiv2.OptimisticLockingStrategy_OPTIMISTIC_LOCKING_STRATEGY_CLIENT,
+				},
+			},
+			state: state{
+				providerTenant: test.DefaultProviderTenant,
+			},
+			wantErr:        true,
+			wantErrMessage: `invalid_argument: optimistic locking is not yet implemented, please do not provide updated_at in update meta`,
+		},
 	}
 
 	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(token.ContextWithToken(t.Context(), tt.sessionToken))
+		t.Run(tt.name, func(innerT *testing.T) {
+			defer testStore.Cleanup(t)
+
+			ctx, cancel := context.WithCancel(token.ContextWithToken(innerT.Context(), tt.sessionToken))
 			defer cancel()
 
-			s := miniredis.RunT(t)
-			c := redis.NewClient(&redis.Options{Addr: s.Addr()})
-
-			tokenStore := token.NewRedisStore(c)
-			certStore := certs.NewRedisStore(&certs.Config{
-				RedisClient: c,
+			test.CreateTenants(innerT, testStore, []*apiv2.TenantServiceCreateRequest{
+				{
+					Name: tt.sessionToken.User,
+				},
 			})
 
-			if tt.tokenToUpdate != nil {
-				err := tokenStore.Set(ctx, tt.tokenToUpdate)
-				require.NoError(t, err)
+			// every logged in users comes with default tenant owner membership
+			test.CreateTenantMemberships(innerT, testStore, tt.sessionToken.User, []*api.TenantMemberCreateRequest{
+				{
+					MemberID: tt.sessionToken.User,
+					Role:     apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			})
+
+			for id, perm := range tt.state.tenantRoles {
+				if id != tt.sessionToken.User {
+					test.CreateTenants(innerT, testStore, []*apiv2.TenantServiceCreateRequest{
+						{
+							Name: id,
+						},
+					})
+				}
+				test.CreateTenantMemberships(innerT, testStore, id, []*api.TenantMemberCreateRequest{
+					{
+						MemberID: tt.sessionToken.User,
+						Role:     perm,
+					},
+				})
 			}
 
-			projectsAndTenantsGetter := func(ctx context.Context, userId string) (*api.ProjectsAndTenants, error) {
-				return &api.ProjectsAndTenants{
-					ProjectRoles: tt.state.projectRoles,
-					TenantRoles:  tt.state.tenantRoles,
-				}, nil
+			for id, perm := range tt.state.projectRoles {
+				test.CreateProjects(innerT, testStore, []*apiv2.ProjectServiceCreateRequest{
+					{
+						Login: tt.sessionToken.User,
+						Name:  id,
+					},
+				})
+				test.CreateProjectMemberships(innerT, testStore, id, []*api.ProjectMemberCreateRequest{
+					{
+						TenantId: tt.sessionToken.User,
+						Role:     perm,
+					},
+				})
 			}
-			log := slog.Default()
-			service := tokenService{
-				log:                      log,
-				tokens:                   tokenStore,
-				certs:                    certStore,
-				issuer:                   "http://test",
-				providerTenant:           tt.state.providerTenant,
-				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
+
+			if tt.tokenToUpdate != nil {
+				err := testStore.GetTokenStore().Set(ctx, tt.tokenToUpdate)
+				require.NoError(innerT, err)
 			}
+
+			service := tokenservice.New(tokenservice.Config{
+				Log:  log,
+				Repo: testStore.Store,
+			})
 
 			if tt.wantErr == false {
 				// Execute proto based validation
@@ -542,15 +2115,22 @@ func Test_Update(t *testing.T) {
 
 			default:
 				got := response.Token
-				assert.Equal(t, tt.wantToken.Uuid, got.Uuid, "uuid")
-				assert.Equal(t, tt.wantToken.Description, got.Description, "description")
-				assert.Equal(t, tt.wantToken.User, got.User, "user id")
-				assert.Equal(t, tt.wantToken.TokenType, got.TokenType, "token type")
-				assert.Equal(t, tt.wantToken.AdminRole, got.AdminRole, "admin role")
-				assert.Equal(t, tt.wantToken.Permissions, got.Permissions, "permissions")
-				assert.Equal(t, tt.wantToken.ProjectRoles, got.ProjectRoles, "project roles")
-				assert.Equal(t, tt.wantToken.TenantRoles, got.TenantRoles, "tenant roles")
-				assert.Equal(t, tt.wantToken.MachineRoles, got.MachineRoles, "machine roles")
+
+				if diff := cmp.Diff(
+					tt.wantToken, got,
+					protocmp.Transform(),
+					protocmp.IgnoreFields(
+						&apiv2.Token{}, "issued_at", "expires",
+					),
+					protocmp.IgnoreFields(
+						&apiv2.Meta{}, "updated_at",
+					),
+				); diff != "" {
+					innerT.Errorf("diff: %s", diff)
+				}
+
+				require.NotNil(innerT, got.Meta.UpdatedAt)
+				assert.True(innerT, !got.Meta.UpdatedAt.AsTime().IsZero())
 			}
 		})
 	}
@@ -558,8 +2138,15 @@ func Test_Update(t *testing.T) {
 
 func Test_Refresh(t *testing.T) {
 	t.Parallel()
+
 	iat := time.Now()
 	exp := iat.Add(time.Hour)
+
+	log := slog.Default()
+
+	testStore, closer := test.StartRepositoryWithCleanup(t, log, test.WithValkey(true), test.WithPostgres(true))
+	defer closer()
+
 	type state struct {
 		providerTenant string
 		projectRoles   map[string]apiv2.ProjectRole
@@ -575,7 +2162,7 @@ func Test_Refresh(t *testing.T) {
 		wantToken      *apiv2.Token
 	}{
 		{
-			name: "can update bare token",
+			name: "can refresh bare token",
 			sessionToken: &apiv2.Token{
 				User:         "phippy",
 				Uuid:         token1,
@@ -592,12 +2179,22 @@ func Test_Refresh(t *testing.T) {
 				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
 				IssuedAt:     timestamppb.New(iat),
 				Expires:      timestamppb.New(exp),
+				Meta: &apiv2.Meta{
+					Generation: 2,
+					CreatedAt:  timestamppb.New(iat),
+					UpdatedAt:  timestamppb.New(iat),
+					Labels: &apiv2.Labels{
+						Labels: map[string]string{
+							"a": "b",
+						},
+					},
+				},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantToken: &apiv2.Token{
-				Uuid:         token1,
+				Uuid:         "something new",
 				User:         "phippy",
 				Permissions:  nil,
 				ProjectRoles: map[string]apiv2.ProjectRole{},
@@ -606,6 +2203,14 @@ func Test_Refresh(t *testing.T) {
 				TokenType:    apiv2.TokenType_TOKEN_TYPE_API,
 				IssuedAt:     timestamppb.New(exp),
 				Expires:      timestamppb.New(exp.Add(time.Hour)),
+				Meta: &apiv2.Meta{
+					Generation: 1,
+					Labels: &apiv2.Labels{
+						Labels: map[string]string{
+							"a": "b",
+						},
+					},
+				},
 			},
 		},
 		{
@@ -634,7 +2239,7 @@ func Test_Refresh(t *testing.T) {
 				Expires:   timestamppb.New(exp),
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 			},
 			wantToken: &apiv2.Token{
 				Uuid:         token1,
@@ -647,6 +2252,9 @@ func Test_Refresh(t *testing.T) {
 				TokenType: apiv2.TokenType_TOKEN_TYPE_API,
 				IssuedAt:  timestamppb.New(exp),
 				Expires:   timestamppb.New(exp.Add(time.Hour)),
+				Meta: &apiv2.Meta{
+					Generation: 1,
+				},
 			},
 		},
 		// FIXME more tests
@@ -660,7 +2268,7 @@ func Test_Refresh(t *testing.T) {
 				},
 			},
 			state: state{
-				providerTenant: "metal-stack",
+				providerTenant: test.DefaultProviderTenant,
 				projectRoles: map[string]apiv2.ProjectRole{
 					kubies: apiv2.ProjectRole_PROJECT_ROLE_EDITOR,
 				},
@@ -672,39 +2280,66 @@ func Test_Refresh(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(token.ContextWithToken(t.Context(), tt.sessionToken))
+		t.Run(tt.name, func(innerT *testing.T) {
+			defer testStore.Cleanup(t)
+
+			ctx, cancel := context.WithCancel(token.ContextWithToken(innerT.Context(), tt.sessionToken))
 			defer cancel()
 
-			s := miniredis.RunT(t)
-			c := redis.NewClient(&redis.Options{Addr: s.Addr()})
-
-			tokenStore := token.NewRedisStore(c)
-			certStore := certs.NewRedisStore(&certs.Config{
-				RedisClient: c,
+			test.CreateTenants(innerT, testStore, []*apiv2.TenantServiceCreateRequest{
+				{
+					Name: tt.sessionToken.User,
+				},
 			})
 
+			// every logged in users comes with default tenant owner membership
+			test.CreateTenantMemberships(innerT, testStore, tt.sessionToken.User, []*api.TenantMemberCreateRequest{
+				{
+					MemberID: tt.sessionToken.User,
+					Role:     apiv2.TenantRole_TENANT_ROLE_OWNER,
+				},
+			})
+
+			for id, perm := range tt.state.tenantRoles {
+				if id != tt.sessionToken.User {
+					test.CreateTenants(innerT, testStore, []*apiv2.TenantServiceCreateRequest{
+						{
+							Name: id,
+						},
+					})
+				}
+				test.CreateTenantMemberships(innerT, testStore, id, []*api.TenantMemberCreateRequest{
+					{
+						MemberID: tt.sessionToken.User,
+						Role:     perm,
+					},
+				})
+			}
+
+			for id, perm := range tt.state.projectRoles {
+				test.CreateProjects(innerT, testStore, []*apiv2.ProjectServiceCreateRequest{
+					{
+						Login: tt.sessionToken.User,
+						Name:  id,
+					},
+				})
+				test.CreateProjectMemberships(innerT, testStore, id, []*api.ProjectMemberCreateRequest{
+					{
+						TenantId: tt.sessionToken.User,
+						Role:     perm,
+					},
+				})
+			}
+
 			if tt.existingToken != nil {
-				err := tokenStore.Set(ctx, tt.existingToken)
+				err := testStore.GetTokenStore().Set(ctx, tt.existingToken)
 				require.NoError(t, err)
 			}
 
-			projectsAndTenantsGetter := func(ctx context.Context, userId string) (*api.ProjectsAndTenants, error) {
-				return &api.ProjectsAndTenants{
-					ProjectRoles: tt.state.projectRoles,
-					TenantRoles:  tt.state.tenantRoles,
-				}, nil
-			}
-			log := slog.Default()
-			service := tokenService{
-				log:                      log,
-				tokens:                   tokenStore,
-				certs:                    certStore,
-				issuer:                   "http://test",
-				providerTenant:           tt.state.providerTenant,
-				projectsAndTenantsGetter: projectsAndTenantsGetter,
-				authorizer:               request.NewAuthorizer(log, projectsAndTenantsGetter),
-			}
+			service := tokenservice.New(tokenservice.Config{
+				Log:  log,
+				Repo: testStore.Store,
+			})
 
 			response, err := service.Refresh(ctx, &apiv2.TokenServiceRefreshRequest{})
 			switch {
@@ -719,14 +2354,21 @@ func Test_Refresh(t *testing.T) {
 
 			default:
 				got := response.Token
-				assert.Equal(t, tt.wantToken.User, got.User, "userId")
-				assert.Equal(t, tt.wantToken.Description, got.Description, "description")
-				assert.Equal(t, tt.wantToken.TokenType, got.TokenType, "token type")
-				assert.Equal(t, tt.wantToken.AdminRole, got.AdminRole, "admin role")
-				assert.Equal(t, tt.wantToken.Permissions, got.Permissions, "permissions")
-				assert.Equal(t, tt.wantToken.ProjectRoles, got.ProjectRoles, "project roles")
-				assert.Equal(t, tt.wantToken.TenantRoles, got.TenantRoles, "tenant roles")
-				assert.Equal(t, tt.wantToken.MachineRoles, got.MachineRoles, "machine roles")
+				if diff := cmp.Diff(
+					tt.wantToken, got,
+					protocmp.Transform(),
+					protocmp.IgnoreFields(
+						&apiv2.Token{}, "uuid", "issued_at", "expires",
+					),
+					protocmp.IgnoreFields(
+						&apiv2.Meta{}, "created_at", "updated_at",
+					),
+				); diff != "" {
+					innerT.Errorf("diff: %s", diff)
+				}
+
+				assert.NotEmpty(innerT, got.Uuid)
+				assert.NotEqual(innerT, tt.existingToken.Uuid, got.Uuid)
 			}
 		})
 	}
